* dxbjavid/dns-extract-long-shift-ub:
  dns: cast remaining ExtractLong byte shifts to uint32_t
  dns: cast to uint32_t in ExtractLong to avoid signed-shift overflow

(cherry picked from commit 60c283b2dcfde697e2052e5e7c0f9bba1d5e5de6)

+++ src/analyzer/protocol/dns/DNS.cc
@@ -621,9 +621,9 @@ uint32_t DNS_Interpreter::ExtractLong(const u_char*& data, int& len) {
 
     uint32_t val;
 
-    val = data[0] << 24;
-    val |= data[1] << 16;
-    val |= data[2] << 8;
+    val = static_cast<uint32_t>(data[0]) << 24;
+    val |= static_cast<uint32_t>(data[1]) << 16;
+    val |= static_cast<uint32_t>(data[2]) << 8;
     val |= data[3];
 
     data += sizeof(val);
